+++ glide/monitors/__init__.py
@@ -1,0 +1,5 @@
+from glide.monitors.classical import ClassicalMeanMonitor
+
+__all__ = [
+    "ClassicalMeanMonitor",
+]